add missing model and service

+++ src/main/java/com/maliitourist/apigestionregions/apigestionregions/controller/PopulationController.java
@@ -70,7 +70,7 @@ public class PopulationController {
     @ApiOperation(value = "Modificatio d'une population.")
     @PutMapping("/misajour/{codePopulation}")
     public ResponseEntity<Object> ModifierRegion(@RequestBody Population population,
-            @PathVariable(value = "codePopulation") int code) {
+            @PathVariable(value = "codePopulation") Long code) {
 
         Population verif_Langue = service.FindByCode(code);
         if (verif_Langue != null) {
@@ -87,7 +87,7 @@ public class PopulationController {
     @ApiOperation(value = "Recuperation d'une population.")
     @GetMapping("/{codePopulation}")
     public ResponseEntity<Object> RecupereLangue(
-            @PathVariable(value = "CodeRegion") int code) {
+            @PathVariable(value = "CodeRegion") Long code) {
 
         Population verif_Langue = service.FindByCode(code);
         if (verif_Langue != null) {
@@ -103,7 +103,7 @@ public class PopulationController {
     @ApiOperation(value = "Su^pression d'une population.")
     @DeleteMapping("/{codePopulation}")
     public ResponseEntity<Object> SuprimerLangue(
-            @PathVariable(value = "CodePopulation") int code) {
+            @PathVariable(value = "CodePopulation") Long code) {
 
         Population verif_Population = service.FindByCode(code);
         if (verif_Population != null) {
