+++ src/main/java/com/maliitourist/apigestionregions/apigestionregions/repository/LangueRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.maliitourist.apigestionregions.apigestionregions.models.Langue;
 
-public interface LangueRepository extends JpaRepository<Langue, Integer> {
+public interface LangueRepository extends JpaRepository<Langue, Long> {
     Langue findByLibele(String nom);
 
     Langue findByCode(String code);
