+++ src/main/java/com/maliitourist/apigestionregions/apigestionregions/repository/PaysRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.maliitourist.apigestionregions.apigestionregions.models.Pays;
 
-public interface PaysRepository extends JpaRepository<Pays, Integer> {
+public interface PaysRepository extends JpaRepository<Pays, Long> {
     Pays findByNom(String nom);
 
     Pays findByCode(String code);
